The operator config and sharing were 4,677 lines in two files
`operator_config.rs` was 2,423: the config itself, the keychain behind
it, validation, applying a change, and the environment the backend is
started with. Six modules now -- `config`, `vault`, `validate`, `apply`,
`backend_env` -- with the guards split the same way.

`sharing.rs` was 2,254: two tunnel providers, a local gateway, the
network interfaces a private share is offered on, and the process
bookkeeping under all of it. Nine modules: `types`, `transitions`,
`ngrok`, `cloudflare`, `gateway`, `interfaces`, `process`, `files`.

Both inherent impls are split across their modules, each carrying the
methods that belong to what it is about.
