+++ desktop/locald/src/sharing/files.rs
@@ -1,0 +1,75 @@
+//! Files only this user may read, and the logs a provider writes.
+
+use super::*;
+
+pub(crate) fn ensure_private_directory(path: &Path) -> io::Result<()> {
+    fs::create_dir_all(path)?;
+    #[cfg(unix)]
+    {
+        use std::os::unix::fs::PermissionsExt;
+        fs::set_permissions(path, fs::Permissions::from_mode(0o700))?;
+    }
+    Ok(())
+}
+
+#[cfg_attr(
+    not(unix),
+    expect(unused_variables, reason = "mode bits are unix-only")
+)]
+pub(crate) fn make_private_file(path: &Path) -> io::Result<()> {
+    #[cfg(unix)]
+    {
+        use std::os::unix::fs::PermissionsExt;
+        fs::set_permissions(path, fs::Permissions::from_mode(0o600))?;
+    }
+    Ok(())
+}
+
+pub(crate) fn bounded_log(path: &Path) -> io::Result<File> {
+    if let Some(parent) = path.parent() {
+        fs::create_dir_all(parent)?;
+    }
+    let truncate = fs::metadata(path)
+        .map(|metadata| metadata.len() > 2 * 1024 * 1024)
+        .unwrap_or(false);
+    let mut options = OpenOptions::new();
+    options.create(true).write(true);
+    if truncate {
+        options.truncate(true);
+    } else {
+        options.append(true);
+    }
+    options.open(path)
+}
+
+pub(crate) fn persist_private_json<T: Serialize>(path: &Path, value: &T) -> io::Result<()> {
+    let contents = serde_json::to_vec_pretty(value).map_err(io::Error::other)?;
+    write_private(path, &contents)
+}
+
+pub(crate) fn write_private(path: &Path, contents: &[u8]) -> io::Result<()> {
+    if let Some(parent) = path.parent() {
+        fs::create_dir_all(parent)?;
+    }
+    let temporary = path.with_extension(format!("{}.next", std::process::id()));
+    let mut options = OpenOptions::new();
+    options.write(true).create(true).truncate(true);
+    #[cfg(unix)]
+    {
+        use std::os::unix::fs::OpenOptionsExt;
+        options.mode(0o600);
+    }
+    {
+        use std::io::Write;
+        let mut file = options.open(&temporary)?;
+        file.write_all(contents)?;
+        file.sync_all()?;
+    }
+    fs::rename(temporary, path)
+}
+
+pub(crate) fn home_dir() -> Option<PathBuf> {
+    std::env::var_os("HOME")
+        .or_else(|| std::env::var_os("USERPROFILE"))
+        .map(PathBuf::from)
+}
